refactor(Account): 당첨금 속성을 int에서 long으로 수정

+++ src/main/java/lotto/domain/Account.java
@@ -11,7 +11,7 @@ public class Account {
     private WinningCalculator winningCalculator;
     private Map<WinningResult, Integer> results = new LinkedHashMap();
     private int count;
-    private int winningAmount;
+    private long winningAmount;
 
     public Account(WinningCalculator winningCalculator) {
         this.winningCalculator = winningCalculator;
